requirements.txt
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dacite==1.8.1
Deprecated==1.2.14
docopt-ng==0.9.0
fastapi==0.103.2
filelock==3.15.4
h11==0.14.0
idna==3.7
jaconv==0.3.4
markdown-it-py==3.0.0
mdurl==0.1.2
mutagen==1.47.0
pathvalidate==3.2.0
platformdirs==4.2.2
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pykakasi==2.3.0
pyTelegramBotAPI==4.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytube==15.0.0
rapidfuzz==3.9.4
redis==5.0.7
requests==2.32.3
rich==13.7.1
scdl==2.11.2
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soundcloud-v2==1.5.4
soupsieve==2.5
spotdl==4.2.5
spotipy==2.24.0
starlette==0.27.0
syncedlyrics==0.7.0
termcolor==2.4.0
text-unidecode==1.3
tqdm==4.66.4
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.2
uvicorn==0.23.2
websockets==12.0
wrapt==1.16.0
yt-dlp==2024.7.16
ytmusicapi==1.7.5
